02_load/pyproject.toml
[project]
name = "processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "click>=8.2.1",
    "dateparser>=1.2.2",
    "isort>=6.0.1",
    "polars>=1.33.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
